Fix JS onReady call
Added jquery-rails and underscore-rails gems to get the assets needed 
for application.js.
Changed the assets in development mode to not be compiled.
Fixed the broken FAQ link on the home page.
Fixed the broken vendor directory include in application.js
Updated the main layout to take a js_module tag from each page so we can
specify what JS gets executed onReady.
Updated the body tag on the FAQ css to match the style of Middleman’s
page_class helper.  We will need to update all other nested pages as 
well.
Added the FAQ page content.
Added the How It Works layout.