Fix sticky navbar.
Moved the content-wrapper div to the default layout so we can add 
uniform padding for domestic navbar to all pages.
Removed unused JS for old signup form.
Added sub-main JS to all pages to control affixed domestic nav.
Updated affixed sidebar on share page to account for new sticky nav 
height.
Updated onReady call to use a new data attribute on the body tag, which 
is just the most specific page_class because the domestic nav adds 
classes to the body tag, so it can’t be relied on.
Added space above all documentation headers so that when they are linked
to it leaves space for the sticky nav.