Initial skeleton for gated download.
Hacked in the login branch of the SW JS SDK from David’s github branch 
to start wiring up the login functionality.
Added some initialization javascript to the Downloads page to trigger a 
login dialogue when the download button is clicked.  Right now, nothing 
happens on the return, but presumably the browser would be redirected to
the download page.  The download page is accessible inside the link DOM,
we can move it to JS if we want to obscure it further, but not a major 
issue.